voice-inbox/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]
openai>=1.0.0
python-dotenv
python-multipart
jinja2
httpx 